+++ Tests/BeforeAll.ps1
@@ -1,3 +1,4 @@
+using namespace Mc2it.Agicap.Authentication
 using namespace System.Diagnostics.CodeAnalysis
 using module ../Agicap.psd1
 
@@ -7,12 +8,12 @@ $clientSecret = ConvertTo-SecureString ($Env:AGICAP_CLIENT_SECRET ?? "BazQux") -
 
 # The client used to query the Agicap API.
 [SuppressMessage("PSUseDeclaredVarsMoreThanAssignments", "client")]
-$client = New-AgicapClient ([pscredential]::new($clientId, $clientSecret))
+$client = New-AgicapClient ([pscredential]::new($clientId, $clientSecret)) -Scope ([Scopes]::All)
 
 # The identifier of the test entity.
 [SuppressMessage("PSUseDeclaredVarsMoreThanAssignments", "entityId")]
 $entityId = [int] ($Env:AGICAP_ENTITY ?? "000000")
 
-# The identifier of the test entity.
+# The identifier of the test organization.
 [SuppressMessage("PSUseDeclaredVarsMoreThanAssignments", "organizationId")]
 $organizationId = [guid]::new($Env:AGICAP_ORGANIZATION ?? "00000000-0000-0000-0000-000000000000")