+++ ChangeLog.md
@@ -2,9 +2,9 @@
 
 ## Version [0.3.0](https://github.com/MC2IT/Agicap.ps1/compare/v0.2.0...v0.3.0)
 - Added the `New-AgicapClient` cmdlet.
-- Organizations: added the `Select-AgicapEntity` and `Select-AgicapOrganization` cmdlets.
-- Payments: added the `New-AgicapBankAccount`, `New-AgicapBeneficiary` and `New-AgicapPostalAddress` cmdlets.
 - Removed the `Get-AgicapApiUrl` cmdlet.
+- **Organizations:** added the `Select-AgicapEntity` and `Select-AgicapOrganization` cmdlets.
+- **Payments:** added the `New-AgicapBankAccount`, `New-AgicapBeneficiary` and `New-AgicapPostalAddress` cmdlets.
 
 ## Version [0.2.0](https://github.com/MC2IT/Agicap.ps1/compare/v0.1.1...v0.2.0)
 - Added the `uriKind` parameter to the `Convert.ToUri()` method.